requirements.txt
pandas
streamlit
plotly
groq
openpyxl
scikit-learn
numpy
smtplib
python-dotenv